requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-dotenv
PyJWT
passlib[bcrypt]
python-jose[cryptography]
